unibind: fold duplicated backend helpers into core render module
The four language backends and gen carried near-verbatim copies of the
render support surface. Fold each to one home:

- RenderError, RenderedInterface, RenderedRecord, name_ident,
  pascal_case, and the Rust boundary-type spelling (rust_type with an
  Ownership mode) now live in unibind_core::render; all four backends
  and their call sites import from there.
- ir::Interface::has_streams replaces the identical predicates in
  backend-ex and gen's ts emitter.
- gen's py emitter reuses unibind-backend-py's stream module (collect,
  StreamExport, class_name) instead of mirroring it, so the stub can
  never drift from the classes the glue registers.
- backend-py and backend-ts render tests ride the test-support helpers
  the ex and jvm tests already use.

Kept apart on purpose: the per-backend tests/fixtures/sample.rs and the
conformance surfaces (deliberately parallel, per their clone:ignore
markers), the conformance build.rs pair (cargo requires a build script
per package and the same four lines exist in a dozen packages repo-wide,
so a unibind-only home would be a second copy, not the home), and the
ts defaults.rs vs py default_tokens pair (different semantics).

Closes #3955

(authored by Claude Code, model Fable 5)
