+++ packages/unibind/backend-ex/src/host.rs
@@ -6,8 +6,9 @@ mod typespec;
 mod wrapper;
 
 use unibind_core::ir;
+use unibind_core::render::RenderError;
 
-use crate::{RenderError, names};
+use crate::names;
 
 /// The rendered Elixir side of the boundary. `unibind-gen`'s `ExEmitter`
 /// decides where the modules land (`lib/<app>/native.ex`, `lib/<app>.ex`).
