+++ packages/unibind/backend-py/Cargo.toml
@@ -16,5 +16,4 @@ unibind-core.workspace = true
 
 [dev-dependencies]
 prettyplease.workspace = true
-serde_json.workspace = true
 unibind-test-support.workspace = true
