+++ packages/unibind/backend-ts/src/lib.rs
@@ -26,38 +26,3 @@ mod stream;
 mod ty;
 
 pub use module::render;
-
-/// The rendered output for one interface.
-pub struct RenderedInterface {
-    /// Sibling items for the exported module: the hidden glue module with
-    /// the `From` error impls, `#[napi]` wrappers, and the stream and
-    /// object handle classes. napi registers exports through link-time
-    /// constructors, so there is no explicit module registration item.
-    pub glue: proc_macro2::TokenStream,
-    /// Attributes to attach to each record struct, index-aligned with the
-    /// interface's records.
-    pub records: Vec<RenderedRecord>,
-}
-
-/// `#[napi(object)]`-shaped attributes for one record struct.
-pub struct RenderedRecord {
-    /// Outer attributes for the struct itself.
-    pub outer: Vec<syn::Attribute>,
-    /// Attributes for each field, index-aligned with the record's fields.
-    pub fields: Vec<Vec<syn::Attribute>>,
-}
-
-/// A rendering failure; the macro positions it at the exported module.
-#[derive(Debug)]
-pub struct RenderError {
-    /// What went wrong and what to do instead.
-    pub message: String,
-}
-
-impl RenderError {
-    pub(crate) fn new(message: impl Into<String>) -> Self {
-        Self {
-            message: message.into(),
-        }
-    }
-}
