+++ packages/unibind/gen/Cargo.toml
@@ -23,6 +23,7 @@ object.workspace = true
 serde_json.workspace = true
 unibind-backend-ex.workspace = true
 unibind-backend-jvm.workspace = true
+unibind-backend-py.workspace = true
 unibind-core.workspace = true
 
 [dev-dependencies]
